Migrate checkbox from /docs/components/checkbox to /components/checkbox
Move all examples (Basic, Terms, Form, Email List) from the old documentation
page to the new reference page, delete the old page, and update all routes,
navigation links, and E2E tests to use the new URL.
